worker/provision_dynamo.py
import boto3
import os
import time
from dotenv import load_dotenv

load_dotenv()

def create_table():
    dynamodb = boto3.resource(
        "dynamodb",
        region_name=os.getenv("AWS_REGION", "us-east-1"),
        aws_access_key_id=os.getenv("AWS_ACCESS_KEY_ID"),
        aws_secret_access_key=os.getenv("AWS_SECRET_ACCESS_KEY")
    )
    
    table_name = "TrendingData"
    
    try:
        print(f"Creating table {table_name}...")
        table = dynamodb.create_table(
            TableName=table_name,
            KeySchema=[
                {'AttributeName': 'source', 'KeyType': 'HASH'}  # Partition key
            ],
            AttributeDefinitions=[
                {'AttributeName': 'source', 'AttributeType': 'S'}
            ],
            ProvisionedThroughput={
                'ReadCapacityUnits': 5,
                'WriteCapacityUnits': 5
            }
        )
        print("Table status:", table.table_status)
        print("Waiting for table to exist...")
        table.wait_until_exists()
        print("✅ Table created successfully!")
        
    except Exception as e:
        print(f"Error creating table: {e}")

if __name__ == "__main__":
    create_table()
