worker/requirements.txt
boto3
schedule
python-dotenv
yt-dlp
pytrends
beautifulsoup4
requests
